requirements.txt
aiohttp~=3.6
PySide2~=5.15
pyyaml~=5.3
